feat(assets): 添加主图配置单PSD文件
新增了主图配置单的PSD设计源文件，用于后续的图片处理和编辑工作。该文件包含了完整的图层结构和设计元素，便于团队成员进行二次创作和修改。
